requirements.txt
qrcode
Pillow
streamlit
